[FEAT] 컬렉션 순서 및 현재 컬렉션 등록 기능 구현
컬렉션 순서 조회 시 순서 1~4가 현재 컬렉션에 해당, 컬렉션 상태 함께 반환

+++ src/main/java/com/nookbook/domain/collection/domain/repository/CollectionRepository.java
@@ -28,9 +28,9 @@ public interface CollectionRepository extends JpaRepository<Collection, Long> {
 
     @Modifying
     @Query("UPDATE Collection c SET c.orderIndex = c.orderIndex - 1 WHERE c.user = :user AND c.orderIndex BETWEEN :start AND :end")
-    void decrementOrderIndexBetween(@Param("user") User user, @Param("start") int start, @Param("end") int end);
+    void decrementOrderIndexBetween(@Param("user") User user, @Param("start") Long start, @Param("end") Long end);
 
     @Modifying
     @Query("UPDATE Collection c SET c.orderIndex = c.orderIndex + 1 WHERE c.user = :user AND c.orderIndex BETWEEN :start AND :end")
-    void incrementOrderIndexBetween(@Param("user") User user, @Param("start") int start, @Param("end") int end);
+    void incrementOrderIndexBetween(@Param("user") User user, @Param("start") Long start, @Param("end") Long end);
 }
